windows: implement new file picker API
