+++ src/os/filesystem_windows.cpp
@@ -24,11 +24,6 @@
 
 #include "filesystem.hpp"
 
-#define HRESULT_TRY(windows_call)                                                                            \
-    if (auto hr = windows_call; !SUCCEEDED(hr)) {                                                            \
-        return FilesystemWin32ErrorCode(HresultToWin32(hr), #windows_call);                                  \
-    }
-
 static constexpr Optional<FilesystemError> TranslateWin32Code(DWORD win32_code) {
     switch (win32_code) {
         case ERROR_FILE_NOT_FOUND:
@@ -1061,226 +1056,6 @@ ErrorCodeOr<Optional<Entry>> Next(Iterator& it, ArenaAllocator& result_arena) {
 
 //
 // ==========================================================================================================
-
-ErrorCodeOr<Span<MutableString>> FilesystemDialogInternal(DialogArguments args) {
-    auto const ids = ({
-        struct Ids {
-            IID rclsid;
-            IID riid;
-        };
-        Ids i;
-        switch (args.type) {
-            case DialogArguments::Type::SaveFile: {
-                i.rclsid = CLSID_FileSaveDialog;
-                i.riid = IID_IFileSaveDialog;
-                break;
-            }
-            case DialogArguments::Type::OpenFile:
-            case DialogArguments::Type::SelectFolder: {
-                i.rclsid = CLSID_FileOpenDialog;
-                i.riid = IID_IFileOpenDialog;
-                break;
-            }
-        }
-        i;
-    });
-
-    IFileDialog* f {};
-    HRESULT_TRY(CoCreateInstance(ids.rclsid, nullptr, CLSCTX_ALL, ids.riid, (void**)&f));
-    ASSERT(f);
-    DEFER { f->Release(); };
-
-    if (args.default_path) {
-        ASSERT(args.default_path->size);
-        ASSERT(IsValidUtf8(*args.default_path));
-        ASSERT(path::IsAbsolute(*args.default_path));
-
-        PathArena temp_path_arena {Malloc::Instance()};
-
-        if (auto const narrow_dir = path::Directory(*args.default_path)) {
-            auto dir = WidenAllocNullTerm(temp_path_arena, *narrow_dir).Value();
-            Replace(dir, L'/', L'\\');
-            IShellItem* item = nullptr;
-            HRESULT_TRY(SHCreateItemFromParsingName(dir.data, nullptr, IID_PPV_ARGS(&item)));
-            ASSERT(item);
-            DEFER { item->Release(); };
-
-            constexpr bool k_forced_default_folder = true;
-            if constexpr (k_forced_default_folder)
-                f->SetFolder(item);
-            else
-                f->SetDefaultFolder(item);
-        }
-
-        if (args.type == DialogArguments::Type::SaveFile) {
-            auto filename = path::Filename(*args.default_path);
-            f->SetFileName(WidenAllocNullTerm(temp_path_arena, filename).Value().data);
-        }
-    }
-
-    if (args.filters.size) {
-        PathArena temp_path_arena {Malloc::Instance()};
-        DynamicArray<COMDLG_FILTERSPEC> win32_filters {temp_path_arena};
-        win32_filters.Reserve(args.filters.size);
-        for (auto filter : args.filters) {
-            dyn::Append(
-                win32_filters,
-                COMDLG_FILTERSPEC {
-                    .pszName = WidenAllocNullTerm(temp_path_arena, filter.description).Value().data,
-                    .pszSpec = WidenAllocNullTerm(temp_path_arena, filter.wildcard_filter).Value().data,
-                });
-        }
-        f->SetFileTypes((UINT)win32_filters.size, win32_filters.data);
-    }
-
-    {
-        PathArena temp_path_arena {Malloc::Instance()};
-        auto wide_title = WidenAllocNullTerm(temp_path_arena, args.title).Value();
-        HRESULT_TRY(f->SetTitle(wide_title.data));
-    }
-
-    {
-        DWORD flags = 0;
-        HRESULT_TRY(f->GetOptions(&flags));
-        HRESULT_TRY(f->SetOptions(flags | FOS_FORCEFILESYSTEM));
-    }
-
-    if (args.type == DialogArguments::Type::SelectFolder) {
-        DWORD flags = 0;
-        HRESULT_TRY(f->GetOptions(&flags));
-        HRESULT_TRY(f->SetOptions(flags | FOS_PICKFOLDERS));
-    }
-
-    auto const multiple_selection = ids.rclsid == CLSID_FileOpenDialog && args.allow_multiple_selection;
-    if (multiple_selection) {
-        DWORD flags = 0;
-        HRESULT_TRY(f->GetOptions(&flags));
-        HRESULT_TRY(f->SetOptions(flags | FOS_ALLOWMULTISELECT));
-    }
-
-    auto const parent_hwnd = (HWND)args.parent_window;
-    if (parent_hwnd) ASSERT(IsWindow(parent_hwnd));
-
-    if (auto hr = f->Show(parent_hwnd); hr != S_OK) {
-        if (hr == HRESULT_FROM_WIN32(ERROR_CANCELLED)) return Span<MutableString> {};
-        return FilesystemWin32ErrorCode(HresultToWin32(hr), "Show()");
-    }
-
-    auto utf8_path_from_shell_item = [&](IShellItem* p_item) -> ErrorCodeOr<MutableString> {
-        PWSTR wide_path = nullptr;
-        HRESULT_TRY(p_item->GetDisplayName(SIGDN_FILESYSPATH, &wide_path));
-        DEFER { CoTaskMemFree(wide_path); };
-
-        auto narrow_path = Narrow(args.allocator, FromNullTerminated(wide_path)).Value();
-        ASSERT(!path::IsDirectorySeparator(Last(narrow_path)));
-        ASSERT(path::IsAbsolute(narrow_path));
-        return narrow_path;
-    };
-
-    if (!multiple_selection) {
-        IShellItem* p_item = nullptr;
-        HRESULT_TRY(f->GetResult(&p_item));
-        DEFER { p_item->Release(); };
-
-        auto span = args.allocator.AllocateExactSizeUninitialised<MutableString>(1);
-        span[0] = TRY(utf8_path_from_shell_item(p_item));
-        return span;
-    } else {
-        IShellItemArray* p_items = nullptr;
-        HRESULT_TRY(((IFileOpenDialog*)f)->GetResults(&p_items));
-        DEFER { p_items->Release(); };
-
-        DWORD count;
-        HRESULT_TRY(p_items->GetCount(&count));
-        auto result = args.allocator.AllocateExactSizeUninitialised<MutableString>(CheckedCast<usize>(count));
-        for (auto const item_index : Range(count)) {
-            IShellItem* p_item = nullptr;
-            HRESULT_TRY(p_items->GetItemAt(item_index, &p_item));
-            DEFER { p_item->Release(); };
-
-            result[item_index] = TRY(utf8_path_from_shell_item(p_item));
-        }
-        return result;
-    }
-}
-
-ErrorCodeOr<Span<MutableString>> FilesystemDialog(DialogArguments args) {
-    // COM initialization is confusing. To help clear things up:
-    // - "Apartment" is a term used in COM to describe a threading isolation model.
-    // - CoInitializeEx sets the apartment model for the calling thread.
-    // - COINIT_APARTMENTTHREADED (0x2) creates a Single-Threaded Apartment (STA):
-    //   - Objects can only be accessed by the thread that created them
-    //   - COM provides message pumping infrastructure (but you still need a message loop)
-    //   - Access from other threads is marshaled through the message queue
-    // - COINIT_MULTITHREADED (0x0) creates a Multi-Threaded Apartment (MTA):
-    //   - Objects can be accessed by any thread in the MTA
-    //   - No automatic message marshaling or pumping
-    //   - Objects must implement their own thread synchronization
-    // - UI components like dialogs require a message pump, so they must be used in an STA.
-    //   Microsoft states:
-    //     "Note: The multi-threaded apartment is intended for use by non-GUI threads. Threads in
-    //     multi-threaded apartments should not perform UI actions. This is because UI threads require a
-    //     message pump, and COM does not pump messages for threads in a multi-threaded apartment."
-    //   By "multi-threaded apartment" they mean COINIT_MULTITHREADED.
-    //
-    // Conclusion: For UI components like IFileDialog, initialize COM with COINIT_APARTMENTTHREADED. If your
-    // thread is already initialized with COINIT_MULTITHREADED, you'll need to create a new thread with
-    // to use UI components.
-    //
-    // As an audio plugin, we can't know for sure the state of COM when we're called.
-
-    if (auto const hr = CoInitializeEx(nullptr, COINIT_APARTMENTTHREADED | COINIT_DISABLE_OLE1DDE);
-        SUCCEEDED(hr)) {
-        // Either we set up COM ourselves, or it was already initialised with the same mode. Either way, we
-        // should call CoUninitialize() later.
-        DEFER { CoUninitialize(); };
-        return FilesystemDialogInternal(args);
-    } else {
-        LogWarning(ModuleName::Filesystem, "{} fallback COM init", __FUNCTION__);
-        // We do not have COINIT_APARTMENTTHREADED, so we must create a new thread with that mode and run the
-        // dialog there.
-        struct Context {
-            DialogArguments const& args;
-            ErrorCodeOr<Span<MutableString>> result;
-        };
-        Context context {.args = args};
-        auto thread = CreateThread(
-            nullptr,
-            0,
-            [](void* p) -> DWORD {
-                auto& context = *(Context*)p;
-
-                auto const hr = CoInitializeEx(nullptr, COINIT_APARTMENTTHREADED | COINIT_DISABLE_OLE1DDE);
-                ASSERT(SUCCEEDED(hr), "new thread couldn't initialize COM");
-                DEFER { CoUninitialize(); };
-
-                context.result = FilesystemDialogInternal(context.args);
-                return 0;
-            },
-            &context,
-            0,
-            nullptr);
-        ASSERT(thread);
-
-        // IFileDialog::Show() will block infinitely because it needs us to handle messages for our
-        // window (WM_SHOWWINDOW for example). We need to pump those messages here.
-        do {
-            MSG msg;
-            while (PeekMessageW(&msg, nullptr, 0, 0, PM_REMOVE)) {
-                LogDebug(ModuleName::Filesystem,
-                         "{} pumping message: 0x{x}, for us: {}",
-                         __FUNCTION__,
-                         msg.message,
-                         msg.hwnd == (HWND)args.parent_window);
-                TranslateMessage(&msg);
-                DispatchMessageW(&msg);
-            }
-        } while (WaitForSingleObject(thread, 10) == WAIT_TIMEOUT);
-
-        CloseHandle(thread);
-        return context.result;
-    }
-}
 
 // Directory watcher
 // Jim Beveridge's excellent blog post on the ReadDirectoryChangesW API:
